Added unit test scheme
